Implement reading `.dat` binary format input files
Also add a new check when reading the input files to make sure the number of the projections available in the file is the same as the program's `N_THETA` value.
Reading `minValue` too now.
Fix typo in `#ifdef` checks.
Now normalize the angle value between 0 (inclusive) and 360 (exclusive).
Update the PGM file reading format.